empty database option added / problems fixed

+++ category-dashboard.php
@@ -6,26 +6,38 @@
 <?php
 	$category_name = "";
 	$category_err  = "";
+	$user_id 	   = $_SESSION['id'];
+	$sql 		   = "SELECT is_admin FROM users WHERE $user_id = users.id";
+	$result 	   = mysqli_query($conn, $sql);
 
-	if($_SERVER["REQUEST_METHOD"] == "POST"){
-		if(!empty($_POST["new_category"])){
-			$category_name = $_POST["new_category"];
-			$sql_request = "INSERT INTO categories(title) VALUES('" . $category_name . "') ";
-
-			if ($conn->query($sql_request) === FALSE) {
-				echo "Error: " . $sql_request . "<br>" . $conn->error;
-			}
+	if ($result->num_rows > 0) {
+		$row = $result->fetch_assoc();
+		if(empty($row)){
+			echo "0 results";
 		}
 	}
-
-	if($_SERVER["REQUEST_METHOD"] == "GET"){
-		if(!empty($_GET['cat_id'])){
-			$delete_category_id = $_GET['cat_id'];
-			$delete_request = "DELETE FROM `categories` WHERE id= " . $delete_category_id . " ";
-		
-				
-			if ($conn->query($delete_request) === FALSE) {
-				echo "Error: " . $delete_request . "<br>" . $conn->error;
+	$is_admin = $row['is_admin'];
+	if($row['is_admin'] == 1){
+		if($_SERVER["REQUEST_METHOD"] == "POST"){
+			if(!empty($_POST["new_category"])){
+				$category_name = $_POST["new_category"];
+				$sql_request = "INSERT INTO categories(title) VALUES('" . $category_name . "') ";
+	
+				if ($conn->query($sql_request) === FALSE) {
+					echo "Error: " . $sql_request . "<br>" . $conn->error;
+				}
+			}
+		}
+	
+		if($_SERVER["REQUEST_METHOD"] == "GET"){
+			if(!empty($_GET['cat_id'])){
+				$delete_category_id = $_GET['cat_id'];
+				$delete_request = "DELETE FROM `categories` WHERE id= " . $delete_category_id . " ";
+			
+					
+				if ($conn->query($delete_request) === FALSE) {
+					echo "Error: " . $delete_request . "<br>" . $conn->error;
+				}
 			}
 		}
 	}
@@ -55,41 +67,44 @@
 										<div class="form-field-wrapper">
 											<input type="text" name="new_category" placeholder="Enter Category Name..."/>
 										</div>
-										<button class="button" >Add New</button>
+										<?php if($row['is_admin'] == 1){?>
+											<button class="button" >Add New</button>
+										<?php }; ?>
 									</div>	
 								</form>
 							</div>
 						</div>
 					</div>
-					<ul class="jobs-listing">
 
-					<?php 
-						$request_category = $conn->query("SELECT * FROM categories ORDER BY title ASC");
+					<?php
+					$request_category = "SELECT * FROM categories ORDER BY title ASC";
+					$page_first_result = ($page-1) * RES_LIMIT;
+					$num_rows = mysqli_num_rows ($conn->query($request_category));
+					$page_total = ceil($num_rows / RES_LIMIT);
+					$request_info = $conn->query($request_category." LIMIT " . $page_first_result . ','. RES_LIMIT);
 
-						while($row = mysqli_fetch_array($request_category, MYSQLI_BOTH)){
-					?>
+					?> <ul class="jobs-listing"> <?php
+					while($row = mysqli_fetch_array($request_info, MYSQLI_BOTH)) { ?>
 						<li class="job-card">
 							<div class="job-primary">
 								<h2 class="job-title"><?php echo $row["title"]?></h2>
 							</div>
 							<div class="job-secondary centered-content">
 								<div class="job-actions">
+									<?php if(isset($is_admin) && $is_admin){?>
 									<a href="<?php echo $_SERVER["PHP_SELF"]?>?cat_id=<?php echo $row['id']; ?>" class="button button-inline">Delete</a>
+									<?php }; ?>
 								</div>
+							</li>
+						<?php
+						}
+						?>
+						<div class="jobs-pagination-wrapper">
+							<div class="nav-links">
+								<?php pagination($page, $page_total); ?>
 							</div>
-						</li>
-					<?php } ?>
-						
-					</ul>					
-					<div class="jobs-pagination-wrapper">
-						<div class="nav-links"> 
-							<a class="page-numbers current">1</a> 
-							<a class="page-numbers">2</a> 
-							<a class="page-numbers">3</a> 
-							<a class="page-numbers">4</a> 
-							<a class="page-numbers">5</a> 
 						</div>
-					</div>
+					</ul>
 				</div>
 			</section>
 		</main>
